channels/journey_bot/requirements.txt
telegram
python-telegram-bot
python-dotenv